requirements.txt
numpy==1.17.3
opencv-python==4.4.0.46
six==1.12.0
tqdm==4.32.2
imageio==2.5.0
gevent==1.4.0
imutils==0.5.3
pyserial==3.4
configobj==5.0.6
whichcraft==0.6.1
Flask==1.1.2
numpy==1.17.3
Pillow==8.0.1
PyAutoGUI==0.9.52
pytesseract==0.3.7
